Moved logic to github_login, set nullable to True
