+++ model/users.py
@@ -6,8 +6,6 @@ import requests
 from itsdangerous import URLSafeTimedSerializer,SignatureExpired
 from mail import mail 
 from flask_mail import Message
-import secrets
-import string
 
 MAILGUN_DOMAIN = "mailgun domain"
 MAILGUN_API_KEY =  "api key"
@@ -21,7 +19,7 @@ class UserModel(db.Model):
     __tablename__="users"
     id = db.Column(db.Integer,primary_key=True)
     username = db.Column(db.String(20),nullable=False,unique=True)
-    password = db.Column(db.String(20),default=UserModel.generate_sample_password())
+    password = db.Column(db.String(20),nullable=True)
     email = db.Column(db.String(40),nullable=False,unique=True)
     activated = db.Column(db.Boolean,default=False)    #set default as False
 
@@ -64,12 +62,6 @@ class UserModel(db.Model):
     @classmethod
     def find_by_email(cls,email):
         return cls.query.filter_by(email=email).first()
-    
-    @classmethod
-    def generate_sample_password(cls):
-        alpha = string.ascii_letters + string.digits
-        random = ''.join(secrets.choice(alpha) for i in range(20))
-        return random
 
     @classmethod
     def check_password(cls,username,password):
