+++ resource/github_login.py
@@ -3,6 +3,8 @@ from flask_restful import Resource
 from outh import github
 from model.users import UserModel
 from flask_jwt_extended import create_access_token,create_refresh_token
+import secrets
+import string
 
 class Github(Resource):
     @classmethod
@@ -28,9 +30,16 @@ class GithubAuthorize(Resource):
         
         #if UserModel.find_by_username(github_username):
         #    return {"msg": "User with username exists"}
-        
+
+    #generate a sample password for github oauth users
+    @classmethod
+    def generate_sample_password(cls):
+        alpha = string.ascii_letters + string.digits
+        random = ''.join(secrets.choice(alpha) for i in range(20))
+        return random
+
         #add user to database
-        user = UserModel(username=github_username,password=None,activated=True,email=github_email)
+        user = UserModel(username=github_username,password=GithubAuthorize.generate_sample_password(),activated=True,email=github_email)
         user.save_to_db()
 
         #create jwt tokens
